Refactor JNI reference tracking  workflow in preparation for new pipeline
The BatchJNIGlobalRefEvents sent from the agent are now stored and queried as-is, and is
up the profilers model to unbtach the samples and look for ref creation/deletion events
tthat satisfy the query range.

The agent-generated MemoryMap protos are also returned as-in as part of the BatchAllocationContexts,
and the address resolution and symblozation steps are handled within the profilers model
instead of before database insertion.

Bug: 121392346
Test: Existing/Updated

Change-Id: I2bca953334f79f6294a4e41d9a27e600c87051f5