First step towards replacing ResourceMergerItem in ResourceFolderRepository
The plan is to use Basic*ResourceItem classes to replace ResourceMergerItem in ResourceFolderRepository and to use RepositoryLoader to do initial loading of the repository. The PSI-related code will remain unchanged for the time being and will be reorganized later.

The resources-aar module will be renamed to resources-base in a separate CL to reflect the fact that the code there is no longer limited to AAR resources.

Git didn't recognize some of the file renames, but all Basic*ResourceItem classes are just renamed Aar*ResourceItem ones. RepositoryLoader and few other related classes were moved to the top level from AarSourceResourceRepository.

Test: existing
Bug: 76409654
Bug: 133779719
Change-Id: If4067fd876f3135f126b05ae78ca8b69b9759f34