[Res Picker] Add DetailedPreview to the ResourcePickerDialog
Adds DetailedPreview to the ResourceExplorerView when it's used for
resource-picking.

This detailed preview includes some information about the
ResourceAssetSet being picked. E.g: may display a preview, basic
metadata like resource name and reference and the configuration/values
availbale for that resource set.

Changed image cache logic for drawables/mipmap/layout, it should now
request a new render for higher resolutions, and does a high quality
scale down for lower resolutions. Uses more memory but reduces the
jankinnes and stutter when changing grid/list view or zomming-in/out,
should also make it possible to eventually share just one instance of
ImageCache for ResourceManager and ResourcePicker (+DetailedPreview).
As it stands, with this CL, there's a chance to use up to 410MB if
ResourceManager and ResourcePicker are both open.

Bug: 133323508
Test: ResourceExplorerViewTest
Change-Id: Iee18b9c1892735c47fb9678f6e6ff23d838149ff