[Gradle Sync] Disable SetSdkDirHyperlinkTest until issue is fixes
This change disables tests that fail due to an issue in the sync
lifecycle, this causes them to hang waiting for source generation to
finish. Note this only happens sometimes.

These tests will be re-enabled once for root cause of this issue is
found and fixed, this issue should only be present in tests and should
not cause any problems for non-test environments.

Bug: 135781629
Test: N/A
Change-Id: I717683f77694dd3ac09049c75295e0d597577c15