Make DataBindingUtil#getBrQualifiedName nullable
DataBindingUtil#getBrQualifiedName relies on getGeneratedPackageName(),
a nullable method in the same class. Currently, if the generated package
name is null, the resulting qualified name will be "null.BR", which is
invalid and causes exceptions to pop up during IntelliJ's class finding.

This CL resolves the issue by making getBrQualifiedName() nullable and
updating its callers to be null-tolerant (e.g. not generating light
BR classes if the qualified name can't be determined).

Test: DataBindingUtilTest still passes.
Bug: 135564124
Change-Id: I2189b1b396f397f64d7e7871529608f94450ed55