Fix isUpToDate in TagToClassMapperImpl (attempt 2)
(This change got reverted because layoutlib tests started failing, but I
believe this was unrelated and has been fixed since).

Before the fix, initial map (with views from all libraries) was always
out of date for modules using >= 20, as it was being built with min API
level of the module but checked against -1. This meant it was being
rebuilt on every keystroke in layout XML.

Bug: 133864394
Test: TagToClassMapperImplTest
Change-Id: I7ca3dbb7cd560ea071ab18ff3e16a3317704565c