[PSD] completion for dependency configuration
Given the recent updates to the QuickSearchComboBox, we can extend its
use in dependency configuration to the PSD Declared Dependency Panel.

The QuickSearchComboBox, being a derivative of the EditorComboBox,
isn't quite a drop-in replacement.  As well as some slight changes to
the logic (creating a new ui element each time rather than re-using
the existing one), we provide some glue to make testing the new UI
component possible: although EditorComboBox subclasses JComboBox, the
difference in implementation details is sufficient that fest-swing
cannot drive it.  Define a minimal Fixture and Driver to handle the
differences in behaviour exposed by existing tests.

Bug: 127954783
Test: existing tests (including UI)
Change-Id: I27f70aac6c2c075c27bdaf0807a7dfc43ceb5365