Remove requireSourceGeneration from syncProject().
The caller should not make assumptions about the underlying project
system implementation details such as what classes are included in
generate sources and therefore cannot know whether the project system
should generate sources or not.

This CL will additionally make generate sources run in few cases where
currently it is skipped. Existing Gradle implementation is supposed to
be fast enough to run on each sync. If it is too slow, there
is/should be a separate option to disable source generation.

Bug: 136160962
Test: existing
Change-Id: If453ea760b33e60abb4deb20cdad416811854f14