Ensure ProfilerDeviceManagerListener subscription
Currently it's possible for ProfilerDeviceManagerListener's subscription
to TransportDeviceManager.TOPIC to occur later than required. When this
happens, ProfilerServiceProxy will not be initialized and profiler
service won't start correctly. This happens because
ProfilerDeviceManagerListener's subscription is registered in
AndroidProfilerToolWindowFactory#createToolWindowContent which is not
guaranteed to be called depending on the tool window's visibility upon
project init.

This CL fixes the issue by registering the subscription in
AndroidProfilerToolWindowFactory#init instead where it will always be
called upon init.

As TransportDeviceManager.TOPIC is a application level topic, the
subscription has been changed to use a application level message bus
instead of a project level message bus.  The FeatureTracker in
ProfilerDeviceManagerListener has been temporarily removed because it
required an instance to the project.

TransportProxy has been made non-final to make mocking possible.

Bug: 132691461
Test: Manually verified that ProfilerDeviceManagerListener is now
always initialized. Added AndroidProfilerToolWindowFactoryTest.
Change-Id: Id75dc4454ed5c6b13541fc952376f297c58c7f8c