Select component in design view based on caret position
When using the split editor, we want to update the selected components in
the preview when we change the caret position in the text editor. Placing
the caret over a component definition in XML should select the component in
the design view. This way it's even more explicit to the user that both
editors are in sync and changes made to one should be reflected in its
counterpart.

Test: UI test to be updated when enabling the flag
Bug: 136174865
Change-Id: Idbc74813bb528a2b073ff832727e8f8e3899e7b2