Fix missing variable assignment in BindingLayoutInfo#update
Without this change, we could generate a lot of unecessary
modifications.

For example, if you go from this:

<data>

to this

<data class="CustomBinding">

and then make other changes to the layout.xml file,
then DefaultBindingLayoutInfo would always think the
data class name kept changing, assuming we were going
from the default name to a custom name each time.

Bug: N/A
Test: Implentation detail only; covered by existing tests
Change-Id: Ibb79770505dfe207d5682d1deaee0052926e68d7