Convert BindingLayoutInfo classes to Kotlin
This collapses some logic with cleaner list comprehensions
and, more importantly, serves to explicitly document fields
as being nullable / not null.

Additionally, fields were reordered with a more consistent
system, and documentation was added. Otherwise, no logic
was changed (intentionally, anyway...)

This CL is a cleanup pass done in preparation for more
refactoring to follow.

Bug: 136500593
Test: Covered by existing tests
Change-Id: If3afbfb1218fe64f81d84467eba156271fa05867