[Dsl] insert new dependencies to preserve order
Arrange for new elements to be inserted at the correct point in the
dependencies block.

This is an approximation in that "the correct point" is not a single
concept.  This implementation uses the order defined by
GradleFileMergers#CONFIGURATION_ORDERING, and places the new
dependency at the first point in the dependencies block where the
current element compares greater (a place preserving the sort when the
dependencies are sorted according to this ordering, and offering some
predictability when the existing dependencies are not sorted).

Bug: 133842504
Test: new Dsl tests and existing ones updated
Test: strengthened psd DependenciesTest uitest
Change-Id: I2236c159249dd14d947dbf2bdab733739f2a7250