Make downloader report which file is in progress now.
As we move SDK downloads to background mode, progress
reporting is limited to the main text only, that appears
in IJ status bar. It is therefore useful to display which
file is being downloaded now, especially given that
download operations are executed sequentially.

Test: Amended existing; to assert progress string
SDK repo ProgressIndicator interface would need to be changed,
which seems an overkill for this change alone.
Bug: N/A

Change-Id: Ifd03a5d41fe135203285a1b37c0123c2792b54c0