[PSD] editable configuration bug fix
UI actions in the PSD declared dependencies view could spuriously
overwrite the configuration of the previously-selected dependency with
the configuration of the new one, as actionPerformed events could fire
following adding the combo box to the panel because of modifications
to its internal state.  Add the actionListener to the combo box after
it has been added to the panel, so that it responds to user events
rather than its own initialization.

Bug: 127954783
Test: added to EditableScopesTest
Change-Id: Ib36d3060c155dc72f60479c6780d8327a65c347f