[Res Explorer] Add support for multi-module search
When entering text in the search field, a LinkLabel will be shown
whenever there are resources that match the search field in other
modules Eg: "4 resources found in module 'A'". It will add a
label for each module that has at least a resource matching the
search criteria.

Clicking the label will change the working module to the one mentioned
in the label.

Bug: 136685114
Bug: 130739013
Test: getOtherModulesResources, showSearchLinkLabels
Change-Id: Iec1cd4a9325d10616595737892392eeec9cd240b