Exclude static fields in DefaultObjectExpander
Class#getDeclaredFields returns both instance and static fields, so
currently there is an edge in the heap graph from every instance of a
class A to each of A's statics (there are also edges from the node for
A's Class object to the statics). This means that if any instance of A
is reachable via a path that does not include A's Class object (which is
likely), none of A's statics (or anything they dominate) will ever be
seen as retained. This is incorrect.

This change filters out the static fields in DefaultObjectExpander, so
that these erroneous edges are not created. This should leave leaktraces
unchanged, as the shortest path to a static field was and is ROOT ->
some ClassLoader -> some Class -> statics. However, it will affect the
listing of retaind objects.

Change-Id: I4e60e8338990a088174ead7e80696086d092c4c7