Enable Go to XML action when using the split editor
Enables the "Go to XML" action when using the split editor to behave the
following way:
  * If we're in design mode, change mode to text and move the caret to the
    desired component.
  * If we're in split mode, preserve the mode and updates the caret position
    similarly to above.

Additionally, this CL restores the current "text + preview" behavior of
updating the cursor of the XML editor, without giving focus to it, when single
clicking an element in the design portion of the split view. Double-clicking
an element (layout-editor only) while in split view should update the cursor
and also give the focus to the text editor.

Bug: 136174865
Test: NlEditorTest#gotoAction UI test should pass with NELE_SPLIT_EDITOR
      enabled or disabled
Change-Id: I016f3585a8a7a2a72cb3ecaae5f2b3b8676b41c6