Introduce data binding scope enlarger for light classes
It turns out our finders were probably returning classes without
considering the search scope correctly - to do this, one really should
call scope.accept(virtualFile). The reason this is a little tricky with
light classes is they don't have files on disk - instead, they are
associated with fake virtual files associated with a view provider. So,
we need to add these fake virtual files to the scope (via the enlarger)
and then check for them in the finders. This ensures we don't return
classes for one module when we restricted our search to another.

Fixes: 126884300
Test: Covered by existing tests
Change-Id: I8ff1bc52bba75c2a4b2465e4ac39b44f86714ae1