Add support for profiler import workflows via EventStreamServer
This CL implements the concept of EventStreamServer which can be
registered with the TransportService as individual streams. It provides
a standard way for users to push non-device specific events and bytes
into the datastore, which can then be queried via the transport service rpcs.

The hprof import workflow for memory profiler has also been implemented
using the EventStreamServer as part of the new pipeline migration. Legacy
allocation tracking and cpu trace imports will be handled via subsequent CLs.

Bug: 136292864
Test: Added

Change-Id: I8562bd69c77db3e3b135ba4fe1d083ad195cbc0b