[Dsl,.kts] resolve extra properties
At the moment, only resolve array lookups where the name of the array
is "extra" and the index is a single, literal string.

This CL also does not properly unquote the text, so variables with
quotes in their name, or other escape characters, won't work
correctly.

Translate the source files for the eleven tests in ExtModelTest for
which this resolution implementation is sufficient.

Bug: 130224064
Test: ExtModelTest.resolveExtProperty for .kts
Test: ExtModelTest.listReferenceInListProperty for .kts
Test: ExtModelTest.mapReferenceInMapProperty for .kts
Test: ExtModelTest.resolveMultiModuleExtProperty for .kts
Test: ExtModelTest.resolveVariableIn{Main,Sub}Module{Build,Property}File for .kts
Test: ExtModelTest.stringReferenceIn{List,Map}Property for .kts
Change-Id: I96be8092aab23e8cc674022825e250ef8dfb3953