Add "import" statements to BR tests
In a separate CL (not yet submitted at this time), I noticed that some
code I changed broke "import (manifest.path).BR" statements silently.
So in this CL, I'm adding tests for this explicitly, and I will rebase
the other CL on top of this one, to ensure it's changes are covered by
tests.

Bug: N/A
Test: Only test code updated
Change-Id: I1a73880ac51c128a63abcda0935d5ddb1d49e9e6