Begin PSI / data separation in BindingLayoutInfo
This is not a minor refactoring, but in general, what is
happening is we are moving from this setup:

          BindingLayoutInfo
                  |
          +-------+-------+
          |               |
DefaultBindingLayoutInfo  |
                   MergedBindingLayoutInfo

to a new setup that both removes inheritance AND separates
PSI-related values from non-PSI, i.e. model, values.

Removing inheritance makes it easier to reason about the
classes, as they have more concrete purposes now, and the
separation of PSI and data values is the first step towards
allowing us to make some performance updates to resource
parsing in the future.

The new class layout is essentionally:

BindingLayoutGroup - represents 1 or more related layouts
BindingLayoutInfo - represents a single layout
BindingLayoutPsi - all PSI related fields, which should be
                   made to load lazily in a future CL
BindingLayoutXml - Data binding values from and of the XML
                   file in data binding layouts

Note that DefaultBindingLayoutInfo and MergedBindingLayoutInfo
classes are now gone. In their place, you simply have a group
of layouts, and one of the layouts is the main (i.e. default)
one.

The main class affected by this refactoring is LightBindingClass,
as it took advantage of the old setup pretty heavily. Now, I've
extracted a `LightBindingClassConfig` class which handles wrapping
the new `BindingLayout` classes and exposing just the values needed
by LightBindingClass.

Bug: 136500593
Test: Covered by existing tests / slight tweaks to existing tests
      in some cases
Change-Id: I63eaa022dcbf8931b0335be1eeb1642ab475538d