Tweak to properly take strings in CreateXmlResourceDialog
Restored 'old' UI, should still properly extract string resources from xml.

Should also allow to type in strings like 'Hello & " \n \t' and save
them as is when the dialog is invoked directly.

Similarly it should be easy to change focus now with just the TAB key
and close (doOkAction) the dialog with the ENTER key.

Bug: 138257967
Test: using "Hello\u0009&amp;😊\n\t&quot;World&quot;\t" to extract from
an XML and typing it as is
Change-Id: Id0fb50ef419f64a2bca9b2b1a73270e6076a07ca