Add perfgate_multi_run tag to perfgate bazel targets
Some of the tests in these bazel targets produce data which is supposed
to be uploaded to perfgate, but currently the data isn't being uploaded
because the bazel targets arent't being run by the perfgate_linux AB
target. Adding the perfgate_multi_run tag will cause the bazel targets
to be run by the perfgate_linux AB target.

Bug: 136658973
Test: existing
Change-Id: Iad07aec1256695f808749e6308ba6e9357232aec