Restart daemon when it's killed (device still online)
This change keeps restarting the daemon in case it's killed.
During the restart, the daemon also reconnects to the agents
that are last known connected.

This change should keep layout inpsector and Echo demo work
when the daemon is killed. Profiler will be fixed in a follow
up CL because it's not using ATTACH_AGENT command directly.

More details on the implementation:

Intuitively, a passive option is to reconnect when the
client (host side) is trying to reach a device via an
rpc call. If no one is actively using the service
anymore, we shouldn't keep the connection alive.

However, relying on rpc calls may not cover our needs
completely. The on-device daemon effectively serves 5
GRPCs in the new pipeline:

GetCurrentTime
GetVersion
GetBytes
Execute
GetEvents

The first four RPCs may be very infrequent. Consider
layout inspector as an example, Execute() is called at
the beginning which contains the app's pid, and
GetEvents() is called after. Because GetEvents() is a
device-wide rpc and no more Execute() would be issued,
the daemon won't be able to know which app it should
talk to by itself. Therefore, we need to keep tracking
connected agents to have a meaningful recover.

There are two ways to detect the daemon is dead
immediately. The first one is the GetEvents()
implementation inside TransportServiceProxy because it
is a streaming rpc. The second one is the
executeShellCommand() in TransportThread::startTransportDaemon().

TransportThread starts the shell command that runs
transport daemon. It makes much sense for it to re-run
the daemon. Therefore, this CL chooses TransportThread
to detect and re-run the daemon, which keeps the
relevant logic concise.

Bug: 135593353
Test: manual. Use Transport Pipeline Echo demo.
Run "adb shell kill TRANSPORT"; the demo should work as
nothing has happend (log would be cleared when agent is
attached/re-attached).

Change-Id: I98ab93c9c21982ceb7e651d1f761b85aab56e448