Refactor legacy allocation tracking workflow to return raw bytes
Previously, the raw bytes are processed in MemoryServiceProxy to generate
the allocation events and contexts queried by the profilers UI. It required
custom synchronization code between stopping a tracking session, waiting for
the byte buffer to be returned and then processed. There were also concerns
around id collision in the class/stack context pool between multiple tracking
sessions.

This change cleans up the workflow by pushing the processing stage to the
LegacyAllocationCaptureObject, so that the context pool is local to each capture.
The logic for MemoryServiceProxy and importing is also significantly simplified
by removing the parsing and synchronization code.

Bug: 121392346
Test: Updated and added
Change-Id: I747e7cf8e49f48acf3fe0750ec2af04baa2a5374