[Dsl,.kts] resolve simple variables and properties
a simple variable is represented in Kotlin Psi by a
KtNameReferenceExpression.  Hooking this into the existing reference
resolution mechanism is straightforward, but note that that existing
mechanism is not correct for Kotlinscript: extra properties and
variable/block properties live in distinct namespaces (accessed by
extra["foo"] and foo respectively) so we should alter our lookup to
reflect this.

There are surprisingly few tests of bare property lookups: most of the
existing tests are for qualified lookups, simple string injections, or
variable lookups.  Nevertheless, one test does exercise this case
directly.

Bug: 130224064
Test: ExtModel.resolveMultiLevelExtProperty
Change-Id: I5ce5e9182744ed54137970a49ac68f7dbb84282b