[Dsl,.kts] injections for list/map dereferences
The theoretically language-agnostic GradleDsl code handles
dereferencing maps and lists, including nested dereferencing: see
GradleDslSimpleExpression.extractElementFromProperties.  It does this
in a not-very-principled way, extracting successive keys with a
regular expression on the text of the expression -- but it so happens
that Kotlin syntax and Groovy syntax are sufficiently close that this
almost works.

Almost?  extra["prop"] also has the syntax of a list/map dereference,
but is mapped in the Gradle Dsl to an ext block.  Dereferences of
extra properties therefore fail in Kotlinscript.

Bug: 130224064
Test: GradlePropertyModel.outOfScopeMapAndListDependencies
Change-Id: I3d92c9a45dfc7866e103b28c387fa288beee0ccd