[Res Explorer] Refactor some logic for rendering resources
There should now be a common place of how resolved values are treated to
be displayed, so that the list on the resource explorer and the summary
view on the resource picker display a resolved value in the same way.

The resources list on the resource explorer should now have a readable value
for the resource types added to AssetPreviewManager.

Bug: 135686734
Test: manually, the resources added to AssetPreviewManager display a
comprehensive string for the resource's value
Change-Id: I9ea87a688667b9972a362d42c2b0d794fc59f416