Add finer-grained Disposer check to BLeak
Track the number of children of a given class for each Disposable,
rather than just the total number of children. This way, if a Disposable
loses a child of type A and gains one of type B, the growth in the
number of type B children will be discovered.

To implement this as part of the main BLeak check, synthetic nodes would
need to be created for each (Disposable, child Class) pair. This would
break the assumption that every node corresponds to a Java object.

Instead, the Disposer status is recorded and analyzed separately, in an
additional phase as part of the end-of-iteration work. Counts of the
number of children of each class for each Disposable are recorded. At
each iteration, only those (Disposable, Class) pairs whose count has
increased are kept.

Result information is now encapsulated in the BleakResult class, as
there are multiple components (the results of the Disposer check and the
main BLeak results).

Change-Id: Id4b522a9147eb205a65388382a71fb8dbe2fceaa