Feature parity for lint quickfixes between Kotlin and Java
Many quickfixes for lint (those implemented on the IDE side by
directly manipulating PSI rather than those described as document
fixes via the LintFix API) were only implemented for Java, not
Kotlin. I added a few additional implementations recently, for
AndroidLintMissingPermissionInspection and for
AndroidLintSwitchIntDefInspection, but this CL more systematically
goes through and implements it for all the remaining lint checks.

Back when the Kotlin plugin depended on the Android plugin instead
of the other way around, quickfixes for lint had to be in the
Kotlin plugin, so there was an extension mechanism for this. This
moves CL all but the ParcelCreator fixes back into the Android
plugin such that they are colocated with the fixes for the other
languages and can share more of the implementation. Some of these
code that I folded in was written in Kotlin so I ported many of
the existing fixes to Kotlin at the same time such that I could
more easily combine them. Since I did this work incrementally one
fix at a time it wasn't practical to prepare before- and after-
ports before applying the additional fixes.

Test: Many new tests, and some are also covered by existing tests
Change-Id: Ifff3022bc27aa66afbcba50373d738a1451190b9