Add buffer to query time range for new pipeline allocation events
The timestamps set on Common.Event in the device daemon can be significantly
delayed (>100ms during peak traffic) from the time when the allocation/deallocation
events actually happened. As a result, even with the +1/-1 samples, the queried
events might not contain all the allocation and deallocation events within the
query range. In those cases, the live allocation visualization logic could fail
due to an incoming deallocation event not having a matching allocation event, and
vice versa.

As a workaround, we include a large enough buffer to ensure we account for all
events within the query range.

Bug: 121392346
Test: Existing

Change-Id: I22e9cc3eb7529a8d53daa1a4fce95c9024c6afe3