Introduce custom AndroidManifestIndex
With the exception of the Manifest Editor, merged manifest
consumers in the IDE only care about a subset of the manifest's
attributes. Currently, there's no way for these consumers to obtain this
information without computing the whole merged manifest, which can take
a very long time depending on the size of the project. This forces
consumers to choose between using a potentially stale view of the
merged manifest (leading to correctness issues) or block on a fresh
view (leading to UI freezes).

To resolve such issues, this CL introduces AndroidManifestIndex,
a custom file-based index which maps each AndroidManifest.xml in the
project to a structured representation of the key attributes from the
tags it contains. We store the attribute values as raw text from the
manifest file because (1) indices are project-independent so placeholder
and resource resolution is unavailable during indexing and (2) even if
we could process the individual manifests further, doing so might
significantly affect indexing time. The index can be disabled using the
new ANDROID_MANIFEST_INDEX_ENABLED flag, which is currently off by
default.

As an example of how we could leverage such an index to perform quick,
accurate merged manifest computations on the fly, this CL also updates
DefaultActivityLocator#computeDefaultActivity to obtain the list of
activities available to a particular module from the AndroidManifestIndex
instead of the actual merged manifest. DefaultActivityLocator will only
use this strategy if DEFAULT_ACTIVITY_LOCATOR_STRATEGY is set to
"INDEX".

Test: DefaultActivityLocatorTest
      + AndroidManifestIndexTest
Bug: 135275364
Change-Id: Ifbb4abd6b470d7a7c7d4d4c6aa921e7440d2890c