[Res Explorer] Add UI feedback for render errors
Sets an error message 'Failed' on the preview for drawables/layouts that
failed to render due to an error in the preview.

Right now the only purpose is to know something went wrong, cause
otherwise it can be hard to tell if that's just how it's supposed to
look. (e.g a drawable of a very faint shadow)

This doesn't cover the case when the preview is innacurate or
imcomplete.

Will eventually do something similar for all resource types at least
when the preview fails for any reason. E.g: the value of a color/font resource is not
resolved.

Example in bug.

Bug: 137867724
Test: n/a
Change-Id: Ic75d4e63fb3a53fc0018682803da189758baf00a