Prototype: Emulator tool window
This CL adds an emulator tool window into the IDE. This toolwindow will
load the emulatorview that is available in emulator builds 5733423 and
later.

This view has very few dependencies, but is far from perfect.

See:
https://android.googlesource.com/platform/external/qemu/+/5797c4143651b3e81702b271878cfb15c13fbbdf

For more details on how it works from the emulator side. A design
document is in the works to address some of the concerns, known issues.

Concerns/Known issues:
   - It uses the telnet protocol to interact with the emulator, you must
   have access to the local filesystem and the auth token needed by the
   emulator.
   - It relies on shared memory, by default on Mac for example not
   enough shared memory is configured. (4mb, we need at least 8mb)
   - Improper keycode translation. There is some rudimentary java
   keycode -> emulator key code translation. Not all keys might work as
   expected on all platforms.
   - No mechanism to select which emulator to display (only 5554 is
   supported)
   - No rotation detection. The current communication channel does not
   support detecting the rotation state.
   - The emulator screen shows the latest frame, even after the emulator
   has crashed (or is restarting).

There is a gRPC bridge that should be used that addresses most of these
concerns.

Test: Emulator windows shows up in linux android-studio build.
Bug: b/138580564
Change-Id: I413345962736dc2d57002ff1e1806c36bba7ee00