Update feature variant only if the variant exists.
Since Gradle model doesn't provide variant dependency on
feature modules, IDE forces the selected variant for feature
module to be the same with base module. However, it is
possible that target variant doesn't exist in feature module.

This CL checks whether the target variant exists or not before
setting it for feature module. If not, keep the current selection
since it is not possible for IDE to correct matching.

Bug: 137289747
Test: testWithNonExistingFeatureVariant
Change-Id: I5a71f90f07714d6c2373222529fa9f815064192b