Do not add renderTemplateValues in project model
Pass information about included formFactors to all models

It was passed from the project model straight to activity model before
which we do not want to do because it complicates the flow.

It may be useful when we have more that 1 model in a new project
(e.g. Wear and Automotive)

Fixes test which was broken by change I0e8161df

Bug: N/A
Test: NewAutomotiveProjectTest
Change-Id: I9d927d1a4a113f4d7f485b561b54e5860aa47aa1