Move DataBindingClassFactory logic to ModuleDataBinding
There was some sketchy lock logic in DataBindingClassFactory,
necessary as it was a series of static functions. Instead, this
logic has been moved to ModuleDataBinding (which
DataBindingClassFactory was delegating to anyway), with the
advantage that we could now create a local lock Object.

After the refactor, DataBindingClassFactory was no longer
necessary, so it has been deleted.

Fixes: 138671228
Test: Covered by existing tests
Change-Id: I909e5f24e864a982c8c416200f1408c62cb87bd4