[Dsl] slightly change one test
In the Groovy DSL, the api() style of declaring dependencies (method
call, rather than application statement) leads to a method with an
argument of type DERIVED, rather than REGULAR.  In the Kotlin DSL, the
only way of declaring the dependencies (the method call) gives a
REGULAR type.

The DERIVED nature of the variable doesn't seem fundamental -- given
that by changing it like in this CL we end up with a different type.
To make it possible to test Kotlin functionality using the same test,
adjust this one and then translate to Kotlin.

Bug: n/a
Test: existing (PropertyDependency.applyFileWithRootDirVariables)
Change-Id: I1a470333ea25367beafa9b8f1919769a447131f1