[Manifest Panel] Remove use of GradleBuildFile and fix hyperlink
This change removes the use of the old GradleBuildFile parser and
replaces it by using the GradleBuildModel. In doing do we also fix the
parsing of the error messages and construction of the hyperlink. Currently
it fails to parse the version of the error message correctly which results
the minSdkVersion being set to a string contain a non-numerical part of the
number.

Bug: N/A
Test: manually tested
Change-Id: I6a1940351c9041f8a2508e4c35ebcee3dc12831c