Update CpuCaptureStage to parse capture on enter
Updated the StatusPane to be a component so it can be used in more
than the L3 of CpuProfilerStage.

Added CpuCaptureHandler to wrap logic behind handing notifications when
a parse is successful vs unsuccessful. This class is also responsible
for handling updating the time it takes to parse a capture. I am using
a range so we don't need yet another aspect for this timing.

Test: Added
Bug: 132268755
Change-Id: Ice7adfe0ad03a6ac6d40df5db8bf01569d816124