[Res Explorer] Fix drag and drop not working on list
Drag and dropping image/svg files in to the resources list wouldn't
bring the ResourceImportDialog since the event was being consume by the
drag handler used to drag drawables/layouts into the LayoutEditor.

The drag handler now delegates the import operation to another object,
in this case, the same DndTarget used for the ResourceImportDialog.

Fixes: 138682200
Fixes: 137363542
Test: manually in linux and mac for both import drawables, and drag and
drop into layout editor
Change-Id: Ib749a01ce746707cb18775c09c25b77b9bcf7add