[Gradle Sync] Stop generating sources on sync
This change prevents sync from triggering a build task to generate
sources after sync. It does not remove the code responsible but
ensures that this task isn't triggered by changing the SyncRequest.

The code that provides this functionality should be removed if it is
decided that we are going to stick with this approach, otherwise this change
should be reverted.

Bug: N/A
Test: existing
Change-Id: I9730d7071a7e3a0214328d77a718a06b8d9ce6e3