Performance hotfix for RenderErrorContributor
This change ensures that the class inheritor search in
RenderErrorContributor yields to write actions, thereby
preventing UI freezes.

The implications of this are that, in some cases, typo suggestion quick
fixes might not appear next to the "missing classes" error in the
layout editor errors panel. I consider this OK because (1) typo
suggestions can already fail to appear for a variety of reasons,
including no close match found or not being in smart mode; (2) UI responsiveness
is much more important than typo suggestions, and (3) the interrupting write
action may cause the render errors to be recomputed again anyway.

There is more work to be done in RenderErrorContributor
(see the linked bug), but this is the safest change that
mitigates the UI freezes seen in 3.5.

Notes:
- Each iteration of the loop in getAllViews() checks
  for cancelation due to the call to psiClass.getQualifiedName()
- The findAll() call in findInheritors checks for
  cancelation as it collects each view.

Bug: 138869517
Test: existing and manual
Change-Id: I5bcf56dfbc3d9794e45ac0e9cd843e0ffe2a885e