Don't use stale merged manifest outside validate()
For TouchBar support, IntelliJ validates run configurations on
the UI thread. This leads to UI freezes in the event that
DefaultActivityLocator#validate needs to block on merged manifest
computation. I595e51f6a attempted to resolve this issue by having
the DefaultActivityLocator use a potentially stale snapshot of the
merged manifest, but it applies this fix too generally, leading to
staleness issues outside of validation where DefaultActivityLocator
returns an outdated or incorrect launch activity name.

This CL limits the use of stale merged manifest snapshots to just
DefaultActivityLocator#validate by default. This means that we will
have to freeze the UI in the event that the merged manifest hasn't
been computed before the user tries to launch the default activity,
so ultimately we will need to fix this freeze by migrating
DefaultActivityLocator to an index-based synchronous API that is
both fast and accurate.

This CL also refactors DefaultActivityLocator to choose a strategy
for obtaining the merged manfiest based on the value of the new
StudioFlags.DEFAULT_ACTIVITY_LOCATOR_STRATEGY string flag and adds
some metrics reporting around how long it actually takes
DefaultActivityLocator to obtain a fresh view of the merged manifest.
These changes make it easier to evaluate our alternatives and switch
between them if they end up introducing severe user-facing issues.

Test: DefaultActivityLocatorTest
Bug: 139421849
Change-Id: Icc43c5d4ebdc64136dbf45f5f29090ae2affb19f