Auto variant selection for dependency modules.
When there are multiple modules share dependency on the same
module, for example,
    app1 -> library1,
    app2 -> library1
if variant selection for app1 and app2 are different, then it
is unavoidable to have conflict on library1. The current code
selects variant for library1 randomly.

With this change, Gradle sync respects the selection of dependent
module whose selection was lastly updated from UI.
For example, if user updated variant of app2 from Build Variant
window, then library1 should be compatible with app2.

Bug: 138404837
Test: testWithSharedDepModules
Change-Id: I92d6b173b89084a8883ba9bd8f39711902a1f7e9