Fix UI tests failing because Gradle tool window was open
UI tests that create a project using the new project wizard were failing
because the Gradle tool window was open after closing the wizard and
would grab the focus, while the test was expecting the editor to have it.

This CL fixes the issue by simply hiding the Gradle tool window when
creating a new project using the wizard.

Bug: 138841171
Bug: 137569703
Test: NavigationViewTest, NewProjectTest, ConstraintLayoutTest
Change-Id: Ib1db7a7e2fb20d86b551f28ecb24655d29dcbefa