Handle environment variable for JDK location
With this CL users can set the JDK Location by setting environment
variable STUDIO_GRADLE_JDK. It is meant to be a temporary setting and as
soon as Studio is launched again without this variable set, the settings
from the PSD should be used.

This cl does not include any changes in the UI, so the values shown in
the PSD does not reflect that the environment variable is used. A future
CL will modify the UI.

Bug: 135292528
Test: Added tests in IdeSdksTest and IdeSdksAndroidTest to check that
the variable is handled correctly.
Change-Id: Iad0048d29a3cfbd4b466134cf8be218ca866c21f