Add new metrics for What's New Assistant
Replace old WNA metrics (AUTO_OPEN/OPEN/CLOSE) with a single event kind
that contains all relevant metrics that we want to log. The
MetricsEventBuilder starts building when the WNA panel is opened, adds
fields as needed when user does certain actions, and then finishes and
logs the actual event when the panel is closed.

Metrics that are logged:
- Whether the panel was auto-opened or manually opened
- Whether the WNA is an "update flow" or regular What's New
- Whether user scrolled to the bottom of the panel (WIP)
- Whether the panel was closed using the Dismiss button (update flow)
- If update flow and user clicked update, how long they took between
  opening panel and updating
- How long the panel was open for before closing for whatever reason

Test: N/A
Bug: 126605387
Change-Id: I5e1ae9af41676ddffbea956add520f892b5cfe48