Fix ComponentAssistantTest UI tests
These tests were trying to drag components to the coordinate (50,50) of
the design surface, but in lower resolutions the design surface width
might be less than 50, causing the unwanted effect of dragging the
component out of the surface bounds. The test then would check for the
presence of such component but would find it.

Bug: 126041973
Test: ComponentAssistantTest
Change-Id: I06d799af6557ab78ed427d6e8050be2d97d8575a