Show STUDIO_GRADLE_JDK in JDK dropdown
... when it is defined as environment variable. The behaviour is
similar to what is done with JAVA_HOME.

The option is meant to help users finding the folder. If
STUDIO_GRADLE_JDK is valid, then it will be used as JDK location
until changed in PSD or Studio restarts.

If the JDK location is changed using the PSD or any other quick fix,
then STUDIO_GRADLE_JDK will no longer be used.

It also adds a notification to let users know when the environment
variable is set incorrectly.

Test: Manually checked changes from PSD are applied as expected.
Bug: 141110179
Change-Id: I08c341bebd1acfe99cf29a45b1f3966086623e2f