[Res Explorer] Fix dragging resources into LayoutEditor
Fixes the cursor not showing the correct state when dragging a resource
into the LayoutEditor or the Xml Editor, it would always show as not
allowed.

Fixes (?) not being able to drag resource into the LayoutEditor in
Windows. Caused from the drag being accepted for an action not listed is
'sourceActions', it would accept a COPY action while the ResourceManager
would expose just LINK action. COPY and MOVE should be the appropiate
actions supported in the DesignerEditor.

Bug: 139046956
Test: Linux and mac for the wrong cursor, TODO on Windows
Change-Id: I4601dc5ae38fbc8dc10a9701dc0565bf012d8f9b