Fix Assistant window showing the bottom upon opening
The purpose of the original hack, detailed in b/77530149, was to reset
the position of the JBScrollBar to the top of the window after the
window had finished rendering. JEditorPane.setText forces the scrollbar
to adjust so that the JEditorPane is visible, so the Assistant window
ends up scrolled to the bottom when opened. The original hack used
SwingUtilities.invokeLater() to set the scrollbar back to the top.

ag/7708549 changed this to Application.invokeLater(), which for whatever
reason does not deterministically reset the scrollbar -- it seems to get
called but then more setText scrolls the window back down again.

DefaultCaret.changeCaretPosition, which is called when the JEditorPane
tries to scroll into view, also uses SwingUtilities.invokeLater(), so it
makes sense to use the same invokeLater here to avoid out-of-order
calls.

Bug: 139094598
Test: N/A
Change-Id: I3c3540b681239d412f7d0c141996e7993466524d