Make SplitEditor use design editor background highlighter
SplitEditor#getBackgroundHighlighter was return the superclass
implementation, which in turn return the text editor's highlighter. The
problem is the design editor also has its own background highlighter,
which is necessary to make the issue panel work.

This CL fixes the issue by making the split editor return a
concatenation of both background highlighters (design editor and text's).

Bug: 139224370
Test: IssuePanelTest#openIssuePanel
Change-Id: Ic9f09e0a06898150a9198b7c572478151d011c1c