Delegate SplitEditor#getFile to its TextEditor
SplitEditor didn't have an implementation of getFile(), so it was
returning FileEditor's default implementation (null). As a result, some of
our listeners that rely on getFile() don't work when opening a file using
the split editor, such as the one in resource manager that checks the file
module to update the resources combobox.

Fixes: 139313284
Test: SplitEditorTest#testFileIsDelegateToTextEditor
Change-Id: I9186fe8254e7867993330a39a1df49cf6e204490