Ask user to confirm Gradle http proxy settings
... when using proxy auto configuration in the IDE.

We currently show a dialog asking users if they want to copy proxy
settings from the IDE to the global gradle.properties file this is done
to prevent further network errors when Gradle tries to download files.
We are doing that only for manual settings.

This CL shows the same dialog when an auto configuration is used. In
this case the initial values are taken from the global gradle.properties
file instead of what the IDE settings are (most likelly they will be
empty on an aut configuration).

Bug: 135102054
Test: testAddProxyConfigureToPropertyFilePAC and existing.
Change-Id: If4b614cfd538e154c73282c5b864a6e5ef2d652e