Move accept() to PreviewElementFinder
Adding a new hasPreviewElements method in PreviewElemenetFinder that
allows the editor provider to delgate the accept to it. For now the
annotation and method implementations use the same approach as before
but we might change them in the future.
Also add TextRanges for the preview elements so they can be located
within the source.

Bug: N/A
Test: Added new regression test to make sure accept works on dumb mode.
Change-Id: I4828196689ef46054b8542d8f5409cd46e1ff9e2