Update finders to work without LocalResourceRepository.getBindingLayoutGroup
Currently, many finders depend on our resource repository logic owning
data binding resources. However, it's questionable they should have
even been put there in the first place; and it doesn't seem to make
sense to add view binding data in on top of that.

Moving forward, we ask the resource repository for raw layout information,
and that, combined with our own data binding specific indices, allows
us to remove our dependency on LocalResourceRepository.getBindingLayoutGroup.

Part of this CL involves introducing a separate fork of
BindingLayoutInfo and BindingLayoutGroup (BindingLayout and
BindingGroup, respectivley) which live in and are only used
by the databinding codebase. This further separates our dependency on
repository code, with the aim of deleting those other classes at
some point in the future.

Fixes: 138406889
Bug: 139468057
Test: Covered by existing

Change-Id: I7d1e14e648459831758266bc4c7e5cb08601c54d